move stringToCarCid to utils
